Added the function to get the match ids
I also removed the apikey from the function get_puuid_with_summoner_name to place it as a global variable. I realized it was needed for every requests. I also cleaned up the code in each function as .json() can be called directly. There's no need for a new variable.